last commit

+++ src/main/java/com/nic/newspaper/dao/UserDAO.java
@@ -9,7 +9,7 @@ public interface UserDAO {
 
 	public List<User> findAll();
 
-	public User findById(int theId);
+	public User findById(long theId);
 
 	public User findByUserName(String theUserName);
